Fix bug in all feature filtering
Summary:
This diff fixes a bug that affected filtering issues with `all_features`, `any_features`, and `exclude_features`. The issue was that we were removing elements from a list while iterating it:

```
            for issue, features in zip(issues, features_list):
                if any_feature_set and not (features & any_feature_set):
                    issues.remove(issue)
                elif all_feature_set and not (
                    features & all_feature_set == all_feature_set
                ):
                    issues.remove(issue)
                elif exclude_feature_set and features & exclude_feature_set:
                    issues.remove(issue)
```
This resulted in the iteration terminating early, and later issues not being properly filtered. The fix was to add a copy

Reviewed By: dkgi

Differential Revision: D22833841

fbshipit-source-id: cbb8441452da952768e5184957dea2014088b99e