refactor taint directory structure to separate out taintAnalysis.ml from the taint library
Summary: The rationale here is that I'm looking to implement ways of specifying models via a query language in the future. TaintAnalysis.ml will need to be able to reference this module, but I'd like the query language bits to live in its own directory for general hygiene. This allows us to reference the taint internals from the query language by breaking the dependency cycle.

Reviewed By: fahndrich

Differential Revision: D22873799

fbshipit-source-id: 84621987866a29cc150a00723524c88b84d3782f