Add argument to __getattr__ stubs
When Python calls a module's `__getattr__` function, it passes it the
name of the attribute. Without including the argument in the signature,
Pyre can't find module attributes not explicitly included in the stubs.
This results in `Undefinied import` errors.

To address this, a `name` argument (typed as `str`) is being added to
each `__getattr__` stub.

Fixes #286