start parsing model query rules
Summary:
This is the next part of the model query DSL work.

We have the types declared in the modelParser module, the idea is the following: we will have a separate module introduced later on whose sole purpose will be to take a list of callables, a list of rules, and create (annotation_kind, taint_annotation) tuples that will be converted to actual models using `add_taint_annotation_to_model`.

Towards this end, the data structure is designed to make that conversion easy, while sharing as much code as possible.

Reviewed By: fahndrich

Differential Revision: D22915978

fbshipit-source-id: 74d27bbc0724db96a43785411ba6f494e0d0abbc