Relativize configuration paths against root
Summary:
Re-apply relativization logic for configuration paths, but do not apply this to `excludes`. Looked into this and I believe this field is the only regex we are passing into the binary, so this shouldn't be a concern for the other values we're fixing here to avoid reliance on changing the working directory.

Will address the excludes field in a later diff.

Reviewed By: pradeep90

Differential Revision: D22912100

fbshipit-source-id: 20003ee7ec68b86c71569010cf4fa5d4f285f0d1