start interpreting parsed query rules for a callable
Summary:
Now that we have the parsed types, this diff adds the ability to apply the parsed rules for a single callable. The output is a list of annotation_kind, taint pairs, which can be fed directly to the model parser to create models.

We take on a regex engine dependency to allow parsing name constraints efficiently.

Reviewed By: fahndrich

Differential Revision: D22941791

fbshipit-source-id: 1ce362d1499781e1f6fab5a28f91854479b4c1c9