Integrated trace operator with interactive.py
Summary:
Added functions for generating traces from trace_operator.py into interactive.py.
The purpose of this was to allow other modules to use the same trace generating tools as interactive.py without having to create an instance of the interactive module. This includes the trace preview feature being developed for SAPP UI

Reviewed By: sinancepel

Differential Revision: D22909951

fbshipit-source-id: 68e8fbfc6104bb6c8d7619e90ec3978cac04c4b0