end-to-end integration of model DSL
Summary: With all the basic pieces in place, this diff puts them all together and starts evaluating model queries. I tried to be a little bit clever about bailing out early to avoid slowing down the system as much as possible, and refactored bits to make sure that we don't duplicate code.

Reviewed By: fahndrich

Differential Revision: D22964314

fbshipit-source-id: e450dce165c49debce95aace87cf612d59ba1bcc