Do not create `SourcePath.t` for non-Python files
Summary:
Turns out that Pyre does not properly exclude files by their extensions in incremental checks. This means that if one touches a file `a.thrift`, we'll treat it as if the person touches `a.py`, and happily run type checking on a file that's obviously non-Python :(

Let's fix the issue by making sure that we stop tracking files whose extensions are not `.py`, `.pyi`, and not included in the `extensions` configuration. Fortunately we have a centralized module `SourcePath` that controls what gets tracked and what doesn't. So the fix is fairly straightforward.

Reviewed By: dkgi

Differential Revision: D22988324

fbshipit-source-id: d0c470adcc4f313325757ee42bbf9378cabdeaf6