Catch SyntaxError during error suppressing
Summary:
It turns out that `ast.check_stable` would raise `UnstableAST` when the transformed source cannot parse, and `SyntaxError` when the original source cannot parse. However, during error suppression we were only catching the `UnstableAST` exception but not the `SyntaxError` exception. This causes `pyre-upgrade` to crash immediately whenever it encounters files that it cannot parse.

Let's fix the issue so that `pyre-upgrade` could avoid the crashing and keep working on the rest of the files that are actually parsable.

Reviewed By: dkgi

Differential Revision: D23036601

fbshipit-source-id: cdf93aa4b97f098eb704efef0eb6a8a811795607