wrap all commands with a try-catch to log error messages
Summary:
Context: When the Command module runs a command which has an uncaught exception, it exits with a code of `1` and swallows the exception, it looks like the following:

```
2020-08-11 13:31:55,880 DEBUG Uncaught exception:
2020-08-11 13:31:55,880 DEBUG
2020-08-11 13:31:55,880 DEBUG   (Unix.Unix_error "No such file or directory" realpath /noexist)
2020-08-11 13:31:55,880 DEBUG
2020-08-11 13:31:55,880 DEBUG Raised at file "command/analyzeCommand.ml", line 186, characters 6-17
2020-08-11 13:31:55,880 DEBUG Called from file "src/command.ml", line 2451, characters 8-238
2020-08-11 13:31:55,880 DEBUG Called from file "src/exn.ml", line 111, characters 6-10
```

This doesn't get transmitted to the client at all due to how we do our logging. If you're interactive mode, you're left with the following, despite the fact that we attempt to catch and log all exceptions in main.ml:

```
pyre
 ƛ Client exited with error code 1
```

This diff takes a blunt instrument approach to the problem since you can't wrap the Command.run - I couldn't figure out a way of creating a Command.spec-like type that would be able to wrap arbitrary functions with try..with blocks, so instead took the easy route.

The end result isn't *amazing*, but it's better than before:

```
[sinancepel@devvm1906]~/local/pytest% pyre analyze
 ƛ Uncaught exception: (Unix.Unix_error "No such file or directory" realpath /noexist)
 ƛ Client exited with error code 1
```

The only exception is the newServer command - it's properly using the `Result` module to avoid this silly problem.

Reviewed By: dkgi

Differential Revision: D23061344

fbshipit-source-id: ec49a75964b72a791686095391028207ba3dca8f