Human readable error when pyre query is run without server
Summary:
In pyre incremental, we check whether the server state is dead and if so, start up a new server. This isn't worthwhile for query because it would involve duplicating a lot of flags related to settings on new servers.

That said, we can at least check this and warn/exit before we hit the message that the json socket file was not found.

Reviewed By: MaggieMoss

Differential Revision: D23114746

fbshipit-source-id: a5b0a7fe90d7d1bf813ab63fd7ac9068a9313acf