New client exit code for server not found
Summary:
Any objections to adding a new class of error code here?

One concern might be that this isn't very useful for other pyre client commands; is relevant in pyre query because failures might come from the fact that a server was not found, or could be a failure from the backend.

Adding this would make it easier for users to automatically detect when this is a problem and kick off a new server, also could make our client exception stats a little more detailed.

Reviewed By: MaggieMoss

Differential Revision: D23115312

fbshipit-source-id: 7800661e942019359be70f7ab3f0c78b98c62847