Bring in source-db buck builder as subclass of BuckBuilder.
Summary:
Problem: The source-db buck builder was being used as a separate binary that would be passed to `--buck-builder-binary`, just like the fast buck builder.

However, given that it's just Python code, we can bring it in as a subclass of `BuckBuilder` and avoid having to dotslash a separate binary.

Reviewed By: grievejia

Differential Revision: D23168663

fbshipit-source-id: 94742a69463fd62ac5b522d04307010479dd608c