Throw invalid configuration consistently on assertions
Summary:
We're doing a few assertions when checking configurations and directly throwing an EnvironmentException; would like to be consistent about throwing Invalid Configuration errors for classification purposes when the problem is something invalid in the config.

Again, this is just for cleaning up the remaining unclassified reliability jobs. Will happily abandon if this conflicts with any configuration cleanup (though support for some consistent prefix for configuration-related errors would be great).

Reviewed By: grievejia

Differential Revision: D23201342

fbshipit-source-id: 83c64a791bfac423cc7b3dac428a5e153dad78f3