Add prefix to invalid configuration error message
Summary:
grievejia I'm happy to abandon this if it conflicts with any of your cleanup changes.

Just got on my radar because the remaining unclassified reliability jobs are mostly invalid configuration related, and I noticed we aren't consistent about prefixing this error in a way I can filter on in the classifier.

Reviewed By: grievejia

Differential Revision: D23201289

fbshipit-source-id: 4fa7964e385afe39aa21fa80bcd4857b2e207f34