Reduce operators on variadics
Summary:
This diff brings support for 2 type operators that given a variadic return a number: Product & Length. We refer to this sort of operators as reduce operators on variadics

This diff is part of a series of diffs to bring arithmetic capabilities at the type level. See previously: D22358132

# Operators
Both operators accept a variadic in any form: ListVariadic, Concatenate, group of types or list of concrete types.

```
Ts = ListVariadic("Ts")
A = TypeVar("A")

def f(...) -> Length[Ts]
def f(...) -> Length[Concatenate[Ts,A]]
def f(...) -> Length[[int,str]]
def f(...) -> Length[int,str]
```

Both operators rely on the existing work on IntExpression, thus Length and Product can interact with other arithmetic operators on types such as Add and Multiply, for example:

```
def f(...) -> Add[Length[Ts], Multiply[Product[Ts], L2]
```

## Length
It returns the length of the variadic once the variadic is replaced by actual type variables.

## Product
Unlike Length, Product only works on subtypes of int. The product rules are the same as for the Multiply operator on types:
 1. Product[..., Any, ...] = Any
 2. Product[..., int, ...] = int
 3. Product[..., Literal[x], Literal[y]] = Product[..., Literal[x * y]]

Moreover, Product takes advantage of arithmetic on IntExpressions to resolve variadics whose typevars have not been fully instantiated:
```
Product[L3,L1,Add[L4,N]] = 3N + 12
```

# Implementation
Length and Product are implemented on top of the existing Polynomial framework to deal with IntExpressions.
For that purpose, we introduce reduce operators at the same level as variables, so that from the point of view of a polynomial, the type variable A and the reduce operator Length[Ts] are treated in the same way, as building blocks of a monomial.
The way how Polynomial works is abstracted with respect to the terms of a monomial, hence we can reuse the existing logic for normalizing the expression. Like variables, operations reduce operators are normalized, for example:

```
Length[Ts] * Length[Ts] = Length[Ts]^2
Length[Ts] + Length[Ts] = 2 * Length[Ts]
````
## Equality
Since reduce operators are also building blocks of Monomials, equality between expressions involving reduce operators is simply the existing equality between polynomials, a monomial by monomial comparison.

# PyTorch Operators
An example where Product is helpful is in torch.flatten (a very common operator):
```
def flatten(t : Tensor[Ts]) -> Tensor[Prod[Ts]]
```
Or view, and derivatives of view, with inference (-1) once we have support for division:
```
def view(self: Tensor[Ts], a: A, b: B, c: Literal[-1])
-> Tensor[A, B, (Prod(Ts) / (A * B)]
```
Regarding Length with its current form I have not identified any torch operator, however there are lots of important operators that could be typed with the more advanced version described in further work.

# Limitations
## Bounds on ListVariadic
Currently ListVariadic does not allow bounds, thus Product cannot specify that it only accepts type variables bounded by int. Currently, it will drop to Any in such case, however once bounds on ListVariadics exist it could be reported as an error.

# Further work
## Converge Multiply and Product
In the future, Multiply and Product could converge into a single operator. Currently, the difference is that Multiply accepts 2 type variables bound by int, while Product accepts any variadic but without bounds. Thus, once that bounds on ListVariadic will be available we could consider merging Product and Multiply.

## Constraint Length
The current proposal could bring multiple enhancements to the type system as a whole. In particular, Length could be used to resonate about the length of variadics and allow the typing of plenty of tensor operators. Example:

```
def mean(x: Tensor[Ts0, M, Ts1], axis: Length[Ts]) -> Tensor[Ts0, Ts1]: ...
```

Currently, when trying to solve L5 <: Length[Ts] we will drop to impossible. However, the type system could be enhanced so that L5 <: Length[Ts] could be resolved by constraining the length of Ts.

Moreover, this sort of operators always implies splitting a variadic into 2 variadics with a type in the middle. However, this is not possible with the current Concatenate, which is limited to a single variadic. Hence, extending Concatenate would also be needed for this purpose.

A list of (some) important operators that could be typed are the following:
 - mean
 - sum
 - argmax
 - argmin
 - prod
 - std
 - std_mean
 - var
 - unsqueeze
 - transpose
 - flatten

## More reduce operators
If needed, with the same architecture we could support other reduce operators. An example could be Sum[]. However, for now it was not implemented since I did not find any use case.

Reviewed By: grievejia

Differential Revision: D22898139

fbshipit-source-id: 00f3aa7d33916fd1f8d5de6321bdc687404f9ebc