Integer Division on Type Variables
Summary:
This diff extends the existing support for arithmetic expression on types with integer division. See previously: D22898139 D22358132

A review on the challenges that supporting division involves and some alternatives for tackling it can be found in https://fb.quip.com/a4fAAhUjqavY

# Operator
Integer division on types is done with the operator "Divide", for example:
```
A = TypeVar("A", bound=int)
B = TypeVar("B", bound=int)

def f(a:A, b:B) -> Divide[A,B]: ...
```

The contexts where Divide is valid is the same as for the operators Add and Multiply. It can be combined with other arithmetic operators without restrictions, and inside a division any valid int expression can appear e.g:
```
def f(a:A, b:B) -> Divide[Add[A,B],Multiply[A,B]] : ..
```

# Semantics
The semantics of `Divide[]` mimic the integer division operator in Python `//`.

## Division by zero
Since division by zero is not defined, when a zero is found in the denominator the resulting type of `Divide[_,0]` will be `Any`, as a way to express that there is not a valid solution for it. It could also be considered in the future to include additional types to detect this sort of illegal behaviors such as `Type(DivisionByZero)`.

## Division of non-literals
When not all the operators of a division are literals, the rules for selecting the return type are the same as in Add and Multiply.

# Equality
One of the main benefits of the existing arithmetic on types is that equality between expression is always decidable. With the introduction of division we keep equality decidable although it introduces some challenges due to the semantics of integer division.
While other operators like Add or Multiply follow a well defined algebra, integer division does not work the same way since many common properties are not satisfied. In particular, adding and multiplying fractions is not always sound, e.g:
```
N/2 + N/2 for N=1
# With addition
N/2 + N/2 = 0
# Without addition
N/2 + N/2 = 2/2 = 1
```
```
N/2 * N/8 for N=7
# With multiplication
N/2 * N/8 = 0
# Without multiplication
N/2 * N/8 = 49/16 -> 3
```
However, there are other operations that are sound and that we need to keep in mind for equality, in particular, fraction simplification:
```
2N/4N = 1/2
```
and grouping:
```
N/2 + N/2 = 2(N/2)
```
Based on this, this proposal makes use of this properties to resonate about equality such that any 2 expressions that are equivalent under this properties are represented in the same way.

A more detailed look at this problem can be found in:
https://fb.quip.com/a4fAAhUjqavY

# Fraction simplification
As exposed in the previous section, fractions are simplified, among other things to allow to resonate about equality. However, for that purpose we need the GCD of multivariate polynomials, what is quite challenging and would involve a lot of extra complexity. For that purpose, we limit ourselves to a smaller set of simplification that should cover the vast majority of use cases. In particular, we obtain the GCD of the constant of the monomials as well as the common variables that may appear, e.g.:

```
2*Length[Ts]*N*M + 4*Length[Ts]*M/ (4*Length[Ts]*M)
->
(N+1)/2
```

# Implementation
A detailed motivation of the implementation can be found in: https://fb.quip.com/a4fAAhUjqavY

We have seen that integer division does not define an algebra and that basically any mutation to the expression tree of a division can lead to unsoundness. Thus, we need a representation that is able to encode division inside polynomials as immutable building blocks while at the same time being able to group and simplify them.

For that purpose, division is implemented following the same schema as in reduce operators on variadics, a division becomes another of the things that a monomial can be made of. Thus, a variable X and a Division P/Q will be treated in the same way, they can be grouped, eliminated, raised to powers, but their internals cannot be mutated till all the variables are resolved.

Moreover, to ensure that division can represent any complex expression, the dividend and quotients of a fraction are polynomials.

## Code organisation
With the introduction of division, now polynomial and monomial are mutually recursive what requires an extra of boilerplate in OCaml to make it work, in particular it requires to define the interface of the modules also in .ml.

# Use cases
Type division will allow to type the integer division operator in Python, as well as a wide variety of ML operators that involve divisions. Currently, In PyTorch I have identified the operators: view, flatten, conv2d, conv3d, maxpool, x[i:j:step], range.

# Other
## Exact division and inverse
One of issues of integer division is that we lose the possibility of simplifying expressions, such `B*(A/B) = A`. Such simplification might be useful in contexts where the developers make sure that the division will have modulo 0, with an assert, if guard, etc. However, that would be "extra" since we anyway need an operator to represent integer division. At this point, it is not very clear how this could be achieved and it is not an issue for the most common use case (convolutions) so if needed this feature could be explored in the future.

## Performance implications
The proposed design for division, and for arithmetic in general, scales constantly with the complexity of the expression. Moreover, it should not affect the performance of any code that does not make use of arithmetic. However, like any other type feature, we cannot prevent the user from building very complex types that may affect performance, (e.g. A/B/C/...). Nonetheless, if we would identify that in large codebases with any operators it slows typechecking we could explore ways of limiting the complexity of expressions.

Reviewed By: grievejia

Differential Revision: D23241424

fbshipit-source-id: 93e50127cdd3e2079e51728f680ec97a1fb30354