Replace `find_local_root` with `find_global_and_local_root`
Summary:
See D23355061 for the motivation to introduce `find_global_and_local_root`. In this diff, we try to replace the custom root finding logic in `pyre-upgrade` with the newly introduced API.

Also, tests for the root finding logic in the upgrade tool relies heavily on mocks. I revamped those to go to the filesystem directly.

Reviewed By: dkgi

Differential Revision: D23355056

fbshipit-source-id: 6fd11be65e46486ee7bc39abb90e8078ea3dc5b1