Replace `find_local_root` with `find_global_and_local_root`
Summary:
See D23355061 for the motivation to introduce `find_global_and_local_root`. In this diff, we try to replace the custom root finding logic in `lsp-adapter` with the newly introduced API.

I wish I could came up with a way to simplify all the directory calculations in the lsp adapter -- but the logic of determining where to put logs/socket files was just too involved for me at the moment. Instead I just did a dumb refactor to preserve the pre-existing semantics there and hope that things could be cleaned up later.

Reviewed By: sinancepel

Differential Revision: D23355058

fbshipit-source-id: 6efd498ad98ac4db16de3e32821359fa367504cb