Introduce an API to find both global and local root
Summary:
Locating the path to both a global and a local configuration root from a given base directory is a fairly common operation in our Python scripts. Yet we do not have a consistent API of doing it before. The fb-internal client, the open-source client, the upgrade tool, and the lsp adapter all seem to craft their own logic, which is less than ideal.

We do have pre-existing APIs for locating global and local roots separately. But the issues are:
- It is always the case that whenever we try to find a local root, we need to find a global root as well. Therefore exposing an API that just find the local root alone is less useful than exposing an API that finds global and local root together.
- The existing API that finds local root actually does a find-global-root internally anyway.

Both of these issues suggest that we'd be better off replacing the API that finds just local root with an API that finds both global and local root. This diff stack attempts to do that and I'm also going to try my best to redirect all the pre-existing custom root-finding logic to the new API to make things more consistent.

This diff is the first step: it introduces the new API `find_global_and_local_root`.

Reviewed By: sinancepel

Differential Revision: D23355061

fbshipit-source-id: cdab016772dd1879f4ff16f7e49acfd7d99ac5a5