Don't crash entire upgrade when line breaks prevent suppression
Summary:
Attempt to add fixmes inside some subset of line break blocks without destroying the syntax and making the file unparseable and therefore unsuppressable.

This does not handle multiline strings/fstrings, but can be expanded to. Holding off bc I'm not sure if this is the direction we want to go in, but I think the only plausible alternative is adding a global suppression (ie., `# pyre-ignore-all-errors[code]`) at the top of the file. Unfortunately, that ruins a lot of other signal in the file, and a reference to a particular line number as rationale for the global ignore very easily gets out of date.

I also could not find a black formatting option that we could pass in to get rid of these line breaks more robustly for us.

At least if this attempt to add an ignore fails because it doesn't handle a given style of line break, it will fail on a file (and just stop trying to suppress) that we would have failed on anyway. Thoughts/suggestions welcome.

Reviewed By: MaggieMoss

Differential Revision: D22787747

fbshipit-source-id: 9597094bc3ea8fe2529350e375f888530676a717