Preserve explicit false options when upgrading configuration.
Summary:
Problem: We recently started tracking the `use_buck_builder` and `use_buck_source_database` fields in the configuration. However, the default behavior of the configuration is to silently drop any fields that are false. Basically, it assumes that `false` is the default value.

However, that is not true for `use_buck_builder`.

Reviewed By: dkgi, grievejia

Differential Revision: D23390386

fbshipit-source-id: 318e7361875b492e17baddb84a1d8e37098909f7