only analyze the part of the call graph that corresponds to the code under analysis
Summary:
Pysa pulls in all the dependencies of projects and analyzes these dependencies in addition to the project under analysis for taint information. However, analyzing all dependencies is wasteful - we only care about the subset of the dependencies that end up getting called by the sources. Thankfully, Pyre already keeps information about whether a source should be type checked, which corresponds exactly to our concept of a project.

Let's only keep the subgraph that is reachable from our sources in order to avoid the analysis of code that doesn't have to do with the project itself.

Context:

Reviewed By: fahndrich

Differential Revision: D23256613

fbshipit-source-id: 3cb82d337808567f0be25b0b58f284553c6612a6