Clean up language to call .pysa files model files
Summary: Pysa models are based on the python stub syntax, so when I originally wrote the doc, I also referred to them as stubs. In reality, they're not true stubs, and we have since standardized on "model" as the name for the stub-like code we put in `.pysa` files. This diff updates the docs to reflect that standardization.

Reviewed By: sinancepel

Differential Revision: D23479203

fbshipit-source-id: 1fb9522461a51e7ff37092fa7cb22be375715f2b