special case lru cache method detection
Summary:
For its callable information, Pysa currently relies on the Callable type containing the name of the underlying callable. This works well enough for standard cases, but doesn't work for decorators like the LRU cache wrapper - lru caches are represented as a callable class, for which Pyre doesn't keep the underlying callable name around.

Since lru caches are very widespread in code, this diff special-cases them and reconstructs the underlying callable to the best of its ability. This is not great, and indicates that we need a better way of getting callable information long term, but it helps with some important cases right now.

To ensure that we're consistent with typeshed, also added a stub integration test.

Reviewed By: dkgi

Differential Revision: D23604625

fbshipit-source-id: ce2b7147296271f615c86178f9ff69890496eb04