Category ServerSecrets returned to user for fbcode
Summary: We currently have "Server secrets reach exit" category defined in opensource which checks flows from ServerSecrets to returnedToUser. I want to have this category to work on fbcode this means adding "ThriftReturn" as a sink. However ThriftReturn is defined in the fbcode taint.config so I have to move its definition to the opensource taint.config

Reviewed By: the-storm

Differential Revision: D23648877

fbshipit-source-id: 8e59263e43569b2a8f4af61665f1e87cac273f84