Preprocess `lazy_import` into import statements
Summary:
lazycatplay

What this diff does: it syntactically replaces this:
```
x = lazy_import("a.b")
```
with this:
```
import a.b as x
```
It also replaces this:
```
x = lazy_import("a.b", "c")
```
with this:
```
from a.b import c as x
```
For now, we only recognize assignments, and the arguments must be string literals for the replacement to happen. We could relax these constraints in the future, but I think what this diff does should cover the majority of the use case.

One notable issue here is how do we recognize if a given function is a lazy import. Options I considered:
1) As long as the function called is named `lazy_import`, we'll treat it as a lazy import. The downside of this approach is that developers who have one of their custom function named  `lazy_import` will be toast.
2) Only recognize `lazy_import` if it is qualified under `libfb.py.lazy_import` module. This would exclude user-defined `lazy_import` which is not actually a lazy import. But the downside is that it leaks the name `libfb.py`, and it does not recognize any open-source lazy imports.
3) Only recognize `lazy_import` if it is qualified under some module, but hide the names of those modules under `Recognized` (which is different for fb-internal and open-source release). This solves the issue in (2) but the downside is that it requires more setup...

Also note that both (2) and (3) have the issue of not being able to follow module aliases. For example,
```
# a.py
from libfb.py.lazy_import import lazy_import
...
# b.py
from a import lazy_import
x = lazy_import("a", "b") # This won't be recognized by (2) and (3) :(
```
Such a limitation is inherent for any pure syntactical transformations. Again I did not find actual use cases so the concern is purely theoretical at the moment.

I did option (2) in the initial revision of this diff for no particular reason so I'm happy to move to (1) and (3) if they are considered better.

Reviewed By: dkgi

Differential Revision: D23633753

fbshipit-source-id: 9444791292f2be137d3dee29b626efb56e4e27c0