Handle line breaks for nodes more generally
Summary:
Factor out as much of the common libcst transformation logic for different nodes as we can.

Slightly unfortunate that where each node holds its trailing whitespace is a bit different, with different keys and also different data structures, so when we're dealing with line breaks over whitespace

meaning:

```
del \
   x
```
or
```
x, y = \
   abc
```

as opposed to line breaks over an expression, like

```
x = a + \
   b
```

we have to modify more than just the single expression node being parenthesized - we have to trim that whitespace (including removing the newline, so this is semantically impt) off of a variable part of the bigger statement node.

This looks different trimming off of Assigns, dels, and other nodes, so that's the bulk of the logic that still lives inside the if-condition of the individual nodes themselves.

Suggestions on making this more condensed/readable welcome. Going to push forward with a few common node types, and then see if we can land this upgrade script and see how it affects our auto land numbers.

Reviewed By: dkgi

Differential Revision: D23255302

fbshipit-source-id: fb3d77a41c916ea2e48203e735d619887f9ee78d