refactor the fetching initial callables bit from the `analyze` function
Summary: The analyze function is the bit that brings it all together - however, right now, it's a bit hard to reason about/manipulate due to everything being inlined; the locals etc. get messy. To help with this, let's start pulling out some logical pieces to the top-level and hopefully demonstrate the data dependencies better.

Reviewed By: dkgi

Differential Revision: D23326691

fbshipit-source-id: 26f83c1623b5aa107cf16b86882077baad5a556a