Filter away type errors on generated getitem call
Summary:
Ran into a bizarre case where `x not in y` check resulted in a parameter error on `__getitem__`.

This was confusing to the user because the error complained that `__getitem__` expected `str` as first param but got `int`, yet `x` was a string.

Glanced at our logic and it looks like we are generating a `__getitem__(0)` call in order to resolve the `In` comparison operator. This is a bit hacky but for now, let's at least omit any errors generated against this "fake" expression.

Note: Added a new test suite bc I didn't see a great place to fit this unit test.

Reviewed By: grievejia

Differential Revision: D23495524

fbshipit-source-id: 2c80172ac549f35df5c89b61fd8c0e983ab9f9cf