Throw error on invalid membership check
Summary: Considered just making the existing Operand error have an optional `left_operand` field but felt this was less self-documenting than adding a new variant, also the two-operand error makes sense as a parameter error but this one is slightly different. Might be considerable to split operand errors to their own class of errors outside the parameter error, but didn't want to spend too much time on this upfront.

Reviewed By: grievejia

Differential Revision: D23713963

fbshipit-source-id: b0cf646e2ef8d161ea1368282050d09e627c230e