add filesystem code smell category
Summary:
Splitting the file system category into 2 categories
* FileSystem
* FileSystem_Codesmell

The filesystem category has a lot of sinks that aren't exploitable directly and make it difficult to productionize. So I am keeping the high signal/impact sinks in the Filesystem and moving the other ones to Codesmell category. This is very similar to what we have been doing in WWW with Zoncolan as well

I kept the codeSmell category for facebook internal configs, not sure if we should open sources these. I don't know where is the limit of opensourcing :) /cc dkgi

Reviewed By: gbleaney

Differential Revision: D23737179

fbshipit-source-id: 9602156ae075ff631fd7c732228d9544303293b1