Reconstruct names for callable protocols
Summary:
Context: We currently do not retain the names of callables for protocols - this presents a challenge for callable classes which we see a lot in practice, where we want Pysa to be able to track them properly in the call graph, etc.

This diff special cases callable protocols in the call graph + call resolution.

Reviewed By: fahndrich

Differential Revision: D23718341

fbshipit-source-id: 4843f9192450a3474cd4e126b7de805406b8253a