Do not throw unbound name errors on defined globals
Summary:
Don't throw unbound name errors if the name is defined as a global. For most builtins, we work around this with a list of hardcoded names in the builtins scope. However, for configerator we'd like to be able to support either shadowing of builtins.pyi or just defining globally accessible names at the stub root.

More context: D23776895

Reviewed By: grievejia

Differential Revision: D23789587

fbshipit-source-id: 26cb28962e84988ead04de3e0388639ab45a985e