Do not try to parse line breaks with no error comments
Summary:
We still don't handle errors inside multiline strings very well, but this is no different from prior support.

Let's make sure we don't unnecessarily parse any python lines that end in a line continuation `\`, but that don't actually have errors to attach.

Pyre upgrade was crashing in distillery due to a `/!\` string which gets parsed into the ast as its own string, since `\` does pick up as line continuation, but then we unnecessarily feed this in to libcst.

Reviewed By: sinancepel, pradeep90

Differential Revision: D23856755

fbshipit-source-id: 7b71f9c13765bb1ebb9caad3875d33ff2b048ce3