Don't leak $unparsed_annotation to error message
Summary:
The synthetic variable `$unparsed_annotation` has been showing up in Pyre error messages. As suggested by the name, this seems to be caused by a failure to parse type annotations.

The synthetic variable was created in `preprocessing.ml`, as there doesn't seem to be a good way to emit a Pyre error from there directly. As the variable is of course usually unbound, it is eventually flagged in `typeCheck.ml`.

This diff just changes the pattern-match in `transform_string_annotation_expression` to yield its input when this sort of parsing failure occurs, rather than generating a synthetic variable. As shown in the test plan, this seems to only remove the previously-generated `$unparsed_annotation` and not produce any new ones.

A more complete solution to this problem would add a new field to the AST to hold information about annotations that failed to parse, and then have `typeCheck.ml` look at that new information and produce a new kind of error. It is not enough to just add a special case to the unbound-name-error code in `typeCheck.ml`, because then the rest of the typechecking still assumes that `$unparsed_annotation` is the type, resulting in more extraneous errors.

Reviewed By: grievejia

Differential Revision: D23848061

fbshipit-source-id: 883edc4d7450fb662af2cc503da34ae387330155