extract first to its own slot for taint domains
Summary:
Context: One of the bottlenecks for functions that do a lot of dictionary manipulations is how we do first index calculations. We look through the set of features to see if a first exists, and only add one if none exists. The idea of this diff is to hoist the First index/field modules out to the top level of the FlowDetails domain, and simplify the representation to be strings.

Note: This makes us lose the `always-` information for first indices. I don't think the idea of `always` plays well with first indices in the first place and have verified that this doesn't cause issues in practice by asking people who are writing filters.

Reviewed By: fahndrich

Differential Revision: D23854058

fbshipit-source-id: f2cacfe43391842eaf7d997766540b7054841e53