Relativize error paths properly from stdin
Summary:
This fixes the last broken client integration test.

Was an annoying error path relativization bug -- turns out when the analysis directory and the directory the user is calling pyre from are different, we run into problems:

- errors coming into the client from the backend have paths relative to the analysis directory
- get errors function parses those errors, then makes sure to relativize them against the directory pyre was called from so they make sense to the user
- when errors are passed in from stdin rather than the backend, it makes sense to expect they will come in relative to the directory the user is calling pyre from when parsing.

This diff makes the distinction - this also ensures that if you call pyre to print out errors and then redirect that output straight back into pyre reading errors from the command line, that paths are handled correctly.

Reviewed By: pradeep90

Differential Revision: D23797622

fbshipit-source-id: 7f72ba75268e9c96e94d3aa4183d3a3c5c115abd